Fix matplotlib exception on Windows.
On Windows, when we make the rqt_plot window very small, it
causes matplotlib to throw an exception like:

numpy.linalg.LinAlgError: Singular matrix

The trouble seems to be that the matplotlib resize event is
trying to invert a zero-sized matrix (I think), so it is
giving numpy fits.  If one of the dimensions of the windows
is 0, just skip trying to apply a tight_layout to it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>