Revert "Add in a pytest dependency for running tests. (#92)"
This reverts commit 6b54799914644a251c5e19868e51dcabc8de7d81.

The main reason to do this is that on Windows Debug, we don't
currently have a Debug wheel for PyQt. Thus these tests always
fail on that platform. While the "correct" solution is to build
a Windows Debug PyQt wheel, that is a decidedly difficult
undertaking. Since we don't have tests to speak of in these
packages, avoid the problem for now by just going back to what
we were doing before, i.e. using unittest for the tests.