Add unit tests for topic name validation & field expansion
Requires #99, #100, #101, #107

This adds unit tests for `rqt_plot.plot_widget.get_plot_fields()`, which
validates the topic name entered by the user and provides a list of
fields/sub-fields to plot.

I thought it was simpler to call `get_plot_fields()` as-is in the tests
and to just ignore the line length linter errors.

Signed-off-by: Christophe Bedard <christophe.bedard@apex.ai>